+++ users/users-model.js
@@ -48,9 +48,6 @@ function getMonthlyBudgetTotal(username) {
     .select('Transportation','Food','HealthInsurance','CarInsurance','HealthLoans','CarLoans','PersonalLoans','Other')
     .first()
 }
-
-
-
 function getMonthlyBudgetByID(id) {
   return db("users")
     .join("monthlybudget", "users.id", "monthlybudget.user_id")
